pyproject.toml
[tool.poetry]
name = "personal-ai-assistant"
version = "0.1.0"
description = ""
authors = ["Chris <ch.grittner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^13.1"
python-dotenv = "^1.0.1"
logging = "^0.4.9.6"
sounddevice = "^0.5.1"
numpy = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
